Add gravity scale modification and ragdoll methods
The modifications introduce methods for modifying the gravity scale and for initiating a ragdoll state. Further, a granular broadcast delegate is set up to propagate gravity direction changes, and an additional method to obtain the character's base primitive component is added. Minor code cleanups and rearrangements have been made as well.